requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
mysql-connector-python==8.0.18
selenium==3.141.0
soupsieve==1.9.5
